Project refactoring - Switched to .NET 6 - Upgraded Avalonia UI to latest release - Moved desktop app into main project - Changed GUI project namespace - Moved project into src directory
