+++ AME.sln
@@ -3,9 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.7.34202.233
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GUI", "GUI\GUI.csproj", "{47539FBA-6856-468F-8A83-66AE40CF88A7}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GUI.Desktop", "GUI.Desktop\GUI.Desktop.csproj", "{6F185286-EFD8-4877-AB9D-8E6FBCD582ED}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GUI", "src\GUI\GUI.csproj", "{47539FBA-6856-468F-8A83-66AE40CF88A7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +15,6 @@ Global
 		{47539FBA-6856-468F-8A83-66AE40CF88A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{47539FBA-6856-468F-8A83-66AE40CF88A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{47539FBA-6856-468F-8A83-66AE40CF88A7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6F185286-EFD8-4877-AB9D-8E6FBCD582ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6F185286-EFD8-4877-AB9D-8E6FBCD582ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6F185286-EFD8-4877-AB9D-8E6FBCD582ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6F185286-EFD8-4877-AB9D-8E6FBCD582ED}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
