+++ Directory.Build.props
@@ -1,6 +1,6 @@
 <Project>
   <PropertyGroup>
     <Nullable>enable</Nullable>
-    <AvaloniaVersion>11.0.2</AvaloniaVersion>
+    <AvaloniaVersion>11.0.5</AvaloniaVersion>
   </PropertyGroup>
 </Project>
