+++ src/GUI/Assets/avalonia-logo.ico
