+++ src/GUI/ViewModels/ViewModelBase.cs
@@ -1,6 +1,6 @@
 ﻿using ReactiveUI;
 
-namespace GUI.ViewModels;
+namespace AME.GUI.ViewModels;
 
 public class ViewModelBase : ReactiveObject
 {
