+++ src/GUI/Views/MainWindow.axaml
@@ -1,12 +1,12 @@
 <Window xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:vm="using:GUI.ViewModels"
+        xmlns:vm="using:AME.GUI.ViewModels"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:views="clr-namespace:GUI.Views"
+        xmlns:views="clr-namespace:AME.GUI.Views"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="GUI.Views.MainWindow"
+        x:Class="AME.GUI.Views.MainWindow"
         Icon="/Assets/avalonia-logo.ico"
-        Title="GUI">
+        Title="AMS2 Mods Editor">
         <views:MainView />
 </Window>
